sorting logic add

+++ src/app/pages/report/report.ts
@@ -2,17 +2,20 @@ import { Component, OnInit, signal, computed } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
+import { NgxPaginationModule } from 'ngx-pagination';
 
 @Component({
   selector: 'app-report',
-  standalone: true, 
-  imports: [FormsModule, CommonModule],
+  standalone: true,
+  imports: [FormsModule, CommonModule, NgxPaginationModule],
   templateUrl: './report.html',
- styleUrls: ['./report.scss']
+  styleUrls: ['./report.scss']
 })
 export class Report implements OnInit {
   reportsList: any[] = [];
   isEditMode: boolean = false;
+  pageNumber: number = 1;
+  itemsPerPage: number = 10;
 
   // Source signals
   // reportsList = signal<any[]>([]);
@@ -49,7 +52,7 @@ export class Report implements OnInit {
     this.http.get('https://api.freeprojectapi.com/api/BusBooking/GetBusVendors').subscribe({
       next: (res: any) => {
         console.log(res);
-        this.reportsList = res;
+        this.reportsList = res.sort((a:any, b:any) => b.vendorId - a.vendorId);
         // this.reportsList.set(res); // set signal value
       },
       error: (err: any) => {
@@ -58,27 +61,11 @@ export class Report implements OnInit {
     })
   }
 
-  // onSave() {
-  //   this.http.post('https://jsonplaceholder.typicode.com/posts', this.reportObj).subscribe({
-  //     next: (res: any) => {
-  //       console.log(res);
-  //       this.reportObj = {
-  //         id: 0,
-  //         userId: 0,
-  //         title: "",
-  //         body: ""
-  //       }
-  //     },
-  //     error: (err: any) => {
-  //       console.log(err);
-  //     }
-  //   })
-  // }
-
   onSave() {
     this.http.post('https://api.freeprojectapi.com/api/BusBooking/PostBusVendor', this.reportObj).subscribe({
       next: (res: any) => {
-        console.log(res);
+        // console.log(res);
+        this.reportsList.unshift(res); // add new report to the top of the list
         this.reportObj = {
           vendorId: 0,
           vendorName: "",
@@ -87,7 +74,7 @@ export class Report implements OnInit {
         }
         this.getReports();
         this.isEditMode = false; // change button text
-        console.log("Report saved successfully!" ,  this.reportObj);
+        console.log("Report saved successfully!", this.reportObj);
       },
       error: (err: any) => {
         console.log(err);
@@ -95,9 +82,9 @@ export class Report implements OnInit {
     })
   }
 
-  onUpdate(){
+  onUpdate() {
     this.http.put('https://api.freeprojectapi.com/api/BusBooking/PutBusVendors?id=' + this.reportObj.vendorId, this.reportObj).subscribe({
-      next:(res : any) => {
+      next: (res: any) => {
         // console.log(res);
         this.getReports();
         this.isEditMode = false;
@@ -108,7 +95,7 @@ export class Report implements OnInit {
           emailId: ""
         }
       },
-      error:(err : any) => {
+      error: (err: any) => {
         console.error("API Error:", err);
       }
     })
@@ -127,17 +114,19 @@ export class Report implements OnInit {
     })
   }
 
-  delete(vendorId: number) { 
-    this.http.delete('https://api.freeprojectapi.com/api/BusBooking/DeleteBusVendor?id=' + vendorId).subscribe({
-      next:(res : any) => {
-        // console.log(res);
-        alert("Report deleted successfully!");
-        this.getReports();
-      },
-      error:(err : any) => {
-        console.error("API Error:", err);
-      }
-    });
+  delete(vendorId: number) {
+    if (confirm("Delete this vendor?")) {
+      this.http.delete('https://api.freeprojectapi.com/api/BusBooking/DeleteBusVendor?id=' + vendorId).subscribe({
+        next: (res: any) => {
+          // console.log(res);
+          // alert("Report deleted successfully!");
+          this.getReports();
+        },
+        error: (err: any) => {
+          console.error("API Error:", err);
+        }
+      });
+    }
   }
 }
